Refactor  GraphQL backend (#964)
