+++ package.json
@@ -102,6 +102,7 @@
     "canvas": "^2.7.0",
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.6",
+    "graphqlviz": "^4.0.1",
     "husky": "^4.3.8",
     "lint-staged": "^10.5.4",
     "node-sass": "^7.0.0",
