+++ package/kedro_viz/api/graphql/__init__.py
@@ -1,0 +1,1 @@
+"""`kedro_viz.api.graphql` defines the GraphQL API."""
