+++ package/kedro_viz/api/rest/__init__.py
@@ -1,0 +1,1 @@
+"""`kedro_viz.api.rest` defines the REST API."""
