+++ package/kedro_viz/data_access/managers.py
@@ -12,7 +12,7 @@ from kedro.pipeline.node import Node as KedroNode
 from sqlalchemy.orm import sessionmaker
 
 from kedro_viz.constants import DEFAULT_REGISTERED_PIPELINE_ID, ROOT_MODULAR_PIPELINE_ID
-from kedro_viz.models.graph import (
+from kedro_viz.models.flowchart import (
     DataNode,
     GraphEdge,
     GraphNode,
@@ -34,6 +34,7 @@ from .repositories import (
     RegisteredPipelinesRepository,
     RunsRepository,
     TagsRepository,
+    TrackingDatasetsRepository,
 )
 
 logger = logging.getLogger(__name__)
@@ -60,17 +61,24 @@ class DataAccessManager:
             lambda: defaultdict(set)
         )
         self.runs = RunsRepository()
+        self.tracking_datasets = TrackingDatasetsRepository()
 
     def set_db_session(self, db_session_class: sessionmaker):
         """Set db session on repositories that need it."""
         self.runs.set_db_session(db_session_class)
 
     def add_catalog(self, catalog: DataCatalog):
-        """Add a catalog to the CatalogRepository.
+        """Add a catalog to the CatalogRepository and relevant tracking datasets to
+        TrackingDatasetRepository.
+
         Args:
             catalog: The DataCatalog instance to add.
         """
         self.catalog.set_catalog(catalog)
+
+        for dataset_name, dataset in self.catalog.as_dict().items():
+            if self.tracking_datasets.is_tracking_dataset(dataset):
+                self.tracking_datasets.add_tracking_dataset(dataset_name, dataset)
 
     def add_pipelines(self, pipelines: Dict[str, KedroPipeline]):
         """Extract objects from all registered pipelines from a Kedro project
