+++ package/kedro_viz/data_access/repositories/__init__.py
@@ -6,3 +6,4 @@ from .modular_pipelines import ModularPipelinesRepository
 from .registered_pipelines import RegisteredPipelinesRepository
 from .runs import RunsRepository
 from .tags import TagsRepository
+from .tracking_datasets import TrackingDatasetsRepository
