+++ package/kedro_viz/server.py
@@ -8,11 +8,12 @@ from kedro.io import DataCatalog
 from kedro.pipeline import Pipeline
 from watchgod import run_process
 
-from kedro_viz.api import apps, responses
+from kedro_viz.api import apps
+from kedro_viz.api.rest.responses import get_default_response
 from kedro_viz.data_access import DataAccessManager, data_access_manager
 from kedro_viz.database import create_db_engine
 from kedro_viz.integrations.kedro import data_loader as kedro_data_loader
-from kedro_viz.models.experiments_tracking import Base
+from kedro_viz.models.experiment_tracking import Base
 
 DEFAULT_HOST = "127.0.0.1"
 DEFAULT_PORT = 4141
@@ -87,8 +88,8 @@ def run_server(
         )
         populate_data(data_access_manager, catalog, pipelines, session_store_location)
         if save_file:
-            res = responses.get_default_response()
-            Path(save_file).write_text(res.json(indent=4, sort_keys=True))
+            response = get_default_response()
+            Path(save_file).write_text(response.json(indent=4, sort_keys=True))
         app = apps.create_api_app_from_project(path, autoreload)
     else:
         app = apps.create_api_app_from_file(load_file)
