+++ package/tests/test_services/test_layers.py
@@ -1,6 +1,6 @@
 import pytest
 
-from kedro_viz.models.graph import GraphNode
+from kedro_viz.models.flowchart import GraphNode
 from kedro_viz.services.layers import sort_layers
