+++ src/apollo/schema.graphql
@@ -1,17 +1,28 @@
-"""Generic scalar type representing a JSON object"""
-scalar JSONObject
+"""
+The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf).
+"""
+scalar JSON @specifiedBy(url: "http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf")
 
 type Mutation {
+  """Update run metadata"""
   updateRunDetails(runId: ID!, runInput: RunInput!): UpdateRunDetailsResponse!
 }
 
 type Query {
+  """Get metadata for specified run_ids from the session store"""
   runMetadata(runIds: [ID!]!): [Run!]!
+
+  """Get metadata for all runs from the session store"""
   runsList: [Run!]!
-  runTrackingData(runIds: [ID!]!, showDiff: Boolean = false): [TrackingDataset!]!
+
+  """Get tracking datasets for specified group and run_ids"""
+  runTrackingData(runIds: [ID!]!, group: TrackingDatasetGroup!, showDiff: Boolean = true): [TrackingDataset!]!
+
+  """Get the installed and latest Kedro-Viz versions"""
   version: Version!
 }
 
+"""Run metadata"""
 type Run {
   author: String
   bookmark: Boolean
@@ -23,6 +34,7 @@ type Run {
   title: String!
 }
 
+"""Input to update run metadata"""
 input RunInput {
   bookmark: Boolean = null
   notes: String = null
@@ -30,26 +42,38 @@ input RunInput {
 }
 
 type Subscription {
+  """Add new runs in real time"""
   runsAdded: [Run!]!
 }
 
+"""Tracking data for a Run"""
 type TrackingDataset {
-  data: JSONObject
-  datasetName: String
-  datasetType: String
+  data: JSON!
+  datasetName: String!
+  datasetType: String!
 }
 
+"""Group to show kind of tracking data"""
+enum TrackingDatasetGroup {
+  METRIC
+  JSON
+}
+
+"""Response for unsuccessful update of run metadata"""
 type UpdateRunDetailsFailure {
   id: ID!
   errorMessage: String!
 }
 
+"""Response for update of run metadata"""
 union UpdateRunDetailsResponse = UpdateRunDetailsSuccess | UpdateRunDetailsFailure
 
+"""Response for successful update of run metadata"""
 type UpdateRunDetailsSuccess {
   run: Run!
 }
 
+"""Installed and latest Kedro-Viz versions"""
 type Version {
   installed: String!
   isOutdated: Boolean!
